% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_rprt.R
\name{utl_jags_rprt}
\alias{utl_jags_rprt}
\title{Wrapper function to produce jags reports for markdown}
\usage{
utl_jags_rprt(x, waic, ...)
}
\arguments{
\item{x}{An object of class rjags (from R2jags) or jagsUI}

\item{waic}{A numeric value for waic}

\item{...}{}
}
\description{
Wrapper function to produce jags reports for markdown
}
