% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_tbl.R
\name{utl_jags_rslt_tbl}
\alias{utl_jags_rslt_tbl}
\title{Pretty jags model results using knitr}
\usage{
utl_jags_rslt_tbl(x)
}
\arguments{
\item{x}{An object of class rjags (from R2jags) or jagsUI}
}
\value{
A knitr::kable with the jags summary
}
\description{
Pretty jags model results using knitr
}
\examples{
jags_result \%>\% utl_jags_rslt_tbl
}
