% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_set.R
\name{utl_jags_set}
\alias{utl_jags_set}
\title{Produce a table of jags model settings from a jags object}
\usage{
utl_jags_set(x, waic = NA, ...)
}
\arguments{
\item{x}{An object of class rjags (from R2jags) or jagsUI}

\item{...}{Another object or value to add to the table}
}
\description{
Produce a table of jags model settings from a jags object
}
\examples{
jags_result \%>\% utl_jags_set(., waic_value)
}
