% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_waic.R
\name{utl_jags_waic}
\alias{utl_jags_waic}
\title{Calcualte WAIC for a jags model using rjags}
\usage{
utl_jags_waic(data, inits, parms, model, chains, burnin, iter, thin)
}
\arguments{
\item{thin}{}
}
\description{
Calcualte WAIC for a jags model using rjags
}
