% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_trc.R
\name{utl_jags_trc}
\alias{utl_jags_trc}
\title{A simple wrapper for MCMC trace to produce jags graphcs in Rmarkdown}
\usage{
utl_jags_trc(x)
}
\arguments{
\item{x}{An object of class rjags (from R2jags) or jagsUI}
}
\value{
returns trace plots of tracked parameters without deviance
}
\description{
A simple wrapper for MCMC trace to produce jags graphcs in Rmarkdown
}
\examples{
jags_result \%>\% util_jags_trc()

}
