% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utl_jags_waic.R
\name{utl_jags_waic}
\alias{utl_jags_waic}
\title{Calcualte WAIC for a jags model using rjags}
\usage{
utl_jags_waic(data, inits, parameters.to.save, model.file, n.chains,
  n.burnin, n.iter, n.thin)
}
\arguments{
\item{inits}{}
}
\description{
Model paremeters follow the same naming conventions as R2jags and jags UI.
}
